+++ UltimateCastbars/Backend/DefaultDB/Player_DB.lua
@@ -756,6 +756,7 @@ UCB.Default_DB.player = {
     },
 
     previewSettings = {
+        showOptions = false,
         previewDuration = 30,
         previewNormalDefaultDuration = false,
         previewEmpowerStages = 5,
@@ -768,6 +769,7 @@ UCB.Default_DB.player = {
         previewNotIntrerruptible = false,
         previewShowKickCD = true,
         previewKickCD = 2,
+        enableMove = true,
     },
 
     copySettings = {